Rearrange modules
It makes sense for the primitives to live in the root module and the internals
to be direct submodules. This also allows internals to use private items of the
primitive types, which is relevant for the `switch` implementation.
